Refactor Parser class and remove unused Test class
